requirements.txt
numpy==1.21.1
Pillow==8.3.1
